+++ platform/src/app/(payload)/admin/importMap.js
@@ -1,3 +1,4 @@
+import { SendInviteButton as SendInviteButton_278df25952fec15c3fd8e010d39bbec6 } from '../../../../src/payload/components/SendInviteButton'
 import { AdminIcon as AdminIcon_5470ea90f7a8fd882c2fe59ff2b1c5b9 } from '../../../../src/payload/components/Logo'
 import { AdminLogo as AdminLogo_5470ea90f7a8fd882c2fe59ff2b1c5b9 } from '../../../../src/payload/components/Logo'
 import { S3ClientUploadHandler as S3ClientUploadHandler_f97aa6c64367fa259c5bc0567239ef24 } from '@payloadcms/storage-s3/client'
@@ -5,6 +6,7 @@ import { CollectionCards as CollectionCards_f9c02e79a4aed9a3924487c0cd4cafb1 } f
 
 /** @type import('payload').ImportMap */
 export const importMap = {
+  "/src/payload/components/SendInviteButton#SendInviteButton": SendInviteButton_278df25952fec15c3fd8e010d39bbec6,
   "/src/payload/components/Logo#AdminIcon": AdminIcon_5470ea90f7a8fd882c2fe59ff2b1c5b9,
   "/src/payload/components/Logo#AdminLogo": AdminLogo_5470ea90f7a8fd882c2fe59ff2b1c5b9,
   "@payloadcms/storage-s3/client#S3ClientUploadHandler": S3ClientUploadHandler_f97aa6c64367fa259c5bc0567239ef24,